Remove --name options from the examples
If we start a container with --name mysql_database and then we stop it, the default Docker behavior is that it will refuse to start a new one until we remove the 'mysql_database'.

It is possible to get pass --cidfile= option which will instruct Docker to write a file with the container id, which may be handy for customers, so let's document it.