Add set of generic and extensible sti scripts
When looking closely at the sti scripts in under
https://github.com/openshift/sti-{php,python,ruby,perl,nodejs}, there are common
scriptlets repeated in all images (duplicit code, that is not exactly the same
in all images).

We should also make the sti scripts to be extensible in upper layer -- i.e. user
takes sti image from OpenShift, wants to extend it a bit, so he shouldn't need
to rewrite whole sti scripts.

This is a proposal to extract the generic part and every image would just include
the specific piece from /usr/share/cont-layer/sti/<hook_name>/. The `<hook_name>`
will be `assemble`, `run` or `usage`. The files that are same in every sti images,
will be added as soon as `sti=true` is set in the rhscl.yaml.

SCL enablement is also not necessary to be done explicitly in the sti scripts, if
SCL is enabled properly and sti scripts are bash scripts.

Since httpd is run in some of the images, this generic extensible way will also
allow to specify running httpd in the same manner.